Loads of small bugfixes for NPEs mostly.
What was happening a lot is that AsyncTasks complete their work and
try to report back on the UI thread, usually by updating some
UI component.

By this time the fragment has already been detached from the Activity
though, causing getActivity() to return null. The code didn't really
take this into account, causing loads of NPEs.

To properly fix this we should probably handle the life-cycle better,
but for now I just checked if the Fragment is still attached in the
relevant places.