Better mechanism to decide between list and detail view.
There are now a number of factors in the decision, mostly based
on the presence of epub and alternate links. If an atom-link
with type alternate exists, we'll assume it to be a details-page
for the entry and load it as such. Else we'll load a fake feed
with a details-screen whenever an epub link is present.