Remember "Scan for Books" dialog settings
When scanning, the choices you made in the dialog are saved for next
time.

Defaults are the same, except for Nook Touch which now defaults to
"/media" for the custom scan path (this is the internal storage, made
available via USB, etc.)

The "Copy to Library" option is now updated when you scan, or when you
change it in the Preferences (changing in the preferences may be
necessary if you only download or Open books rather than scanning for
them).