Standardise indentation, closes #573, add style note to README
This was done via macros but it looks OK to me at a once-over, it should
be possible to change anything that seems wrong.

Indents are 4 spaces, as requested. All tabs removed.

Multiline argument lists are indented two levels beneath the function
definition or call.

Opening braces haven't been moved to the same line as the defining
function if they weren't already there. There seem to be relatively few
of these anyhow.

For the record, the output of this is emacs' standard java-mode
indentation with the following hook set:

> ; Don't indent argument lists aggressively in Java
> (add-hook 'java-mode-hook '(lambda ()
>     (c-set-offset 'arglist-intro '+)
>          (c-set-offset 'arglist-cont-nonempty '++)
>	       ))

... which is hopefully fairly similar to what IntelliJ does.